+++ app/templates/karma.conf.js
@@ -55,26 +55,17 @@ module.exports = function(config) {
 
         // start these browsers
         // available browser launchers: https://npmjs.org/browse/keyword/karma-launcher
-        browsers: ['Chrome'],
-
-        customLaunchers: {
-            /*eslint-disable */
-            Chrome_travis_ci: {
-                base: 'Chrome',
-                flags: ['--no-sandbox']
-            }
-        },
+        browsers: ['PhantomJS'],
 
         // Continuous Integration mode
         // if true, Karma captures browsers, runs the tests and exits
         singleRun: false
     };
 
-    if(process.env.TRAVIS){
-        configuration.browsers = ['Chrome_travis_ci'],
+    if (process.env.TRAVIS) {
         configuration.coverageReporter = {
-            type : 'lcov',
-            dir : 'coverage/'
+            type: 'lcov',
+            dir: 'coverage/'
         },
         configuration.singleRun = true;
     }